+++ CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.0] (2021-10-13)
+
+### Added
+
+- Added support for new types: AnyUrl, HttpAnyUrl, HttpUrl, StrictUrl, bool and StrictBool.
+
+### Changed
+
+- Changed development status to alpha.
+
 ## [0.2.1] (2021-10-03)
 
 ### Fixed
 
-- Fix `NOT NULL` constraint behavior.
+- Fix `NOT NULL` behavior.
 
 ### Changed
 
@@ -30,6 +40,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First Release.
 
-[0.2.1]: https://github.com/dcruzf/validatable/compare/v0.1.0-alpha...v0.2.1
+[0.3.0]: https://github.com/dcruzf/validatable/compare/v0.2.1...v0.3.0
+[0.2.1]: https://github.com/dcruzf/validatable/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/dcruzf/validatable/compare/v0.1.0-alpha...v0.2.0
 [0.1.0]: https://github.com/dcruzf/validatable/releases/tag/v0.1.0-alpha